+++ packages/trpc/server/routers/viewer/apiKeys/list.handler.ts
@@ -1,4 +1,4 @@
-import { ApiKeyRepository } from "@calcom/lib/server/repository/apiKey";
+import { ApiKeyRepository } from "@calcom/lib/server/repository/apiKeyRepository";
 
 import type { TrpcSessionUser } from "../../../types";
